// src/Types/form.ts
export interface SubmissionForm {
    dateOfIssue?: string;
    candidateName?: string;
    jobTitle?: string;
    reportingTo?: string;
    startDate?: string;
    employmentType?: string;
    annualCTA?: string;
    workScheduleStart?: string;
    workScheduleEnd?: string;
    jobLocation?: string;
}